% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{c.term}
\alias{c.term}
\title{Combine Term Vectors}
\usage{
\method{c}{term}(..., recursive = FALSE)
}
\arguments{
\item{...}{Term vectors to concatenate}

\item{recursive}{Ignored.}
}
\value{
A term vector.
}
\description{
Combine Term Vectors
}
\examples{
c(NA_term_, "b", "a[1]", "")
}
\seealso{
\code{\link[=term-vector]{term-vector()}}
}
