% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateDTE.R
\name{calculateDTE}
\alias{calculateDTE}
\title{Calculate the differential expression of mRNAs and RPFs, and the differential translation of genes.}
\usage{
calculateDTE(raw_rna, raw_ribo, rna_label, ribo_label, baseLevel,
  minCounts = 10, minCountsProportion = 1)
}
\arguments{
\item{raw_rna}{raw RNA-seq read counts}

\item{raw_ribo}{raw Ribo-seq read counts}

\item{rna_label}{the experiment conditions corresponding to the samples in raw_rna}

\item{ribo_label}{the experiment conditions corresponding to the samples in raw_ribo}

\item{baseLevel}{the condition used as the base level for comparisons of differential expression and differential translation}

\item{minCounts}{the number of minimum read count of a gene in a sample, with default value 10}

\item{minCountsProportion}{the proportion of samples where a gene exhibits read counts larger than the minCounts in a same condition, with default value 1}
}
\value{
a dataframe consists of DESeq2 results of mRNA and RPF differential expression and Riborex results of differential translation
}
\description{
Calculate the differential expression of mRNAs and RPFs, and the differential translation of genes.
}
\examples{
calculateDTE()
calculateDTE(raw_rna, raw_ribo, rna_label, ribo_label, baseLevel, minCounts=10, minCountsProportion=1 )
}
