% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_term}
\alias{check_term}
\title{Check Term Vector}
\usage{
check_term(x, nas = TRUE, length = NA, unique = FALSE,
  sorted = FALSE, named = NA, valid = FALSE, complete = FALSE,
  attributes = named, names = TRUE, x_name = substitute(x),
  error = TRUE)
}
\arguments{
\item{x}{The object to check.}

\item{nas}{A logical scalar specifying whether x can include missing values.}

\item{length}{A flag indicating whether x should have elements (versus no elements) or a missing value indicating no requirements or a count or count range of the number of elements or a count vector of the permitted number of elements.}

\item{unique}{A flag indicating whether the values must be unique.}

\item{sorted}{A flag indicating whether the vector must be sorted.}

\item{named}{A flag indicating whether the vector must be named or unnamed or NA if it doesn't matter.}

\item{valid}{A logical scalar specifying whether x must not require repairing.}

\item{complete}{A logical scalar specifying whether x must be complete.}

\item{attributes}{A flag indicating whether the vector must or must not have attributes or NA if it doesn't matter.}

\item{names}{A flag specifying whether names should be considered an attribute.}

\item{x_name}{A string of the name of the object.}

\item{error}{A flag indicating whether to throw an informative error or immediately generate an informative message if the check fails.}
}
\value{
An invisible copy of x (if it doesn't throw an error).
}
\description{
Check term vector.
}
\examples{
check_term(as.term("a"))
}
