% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk.R
\name{chk_pars}
\alias{chk_pars}
\title{Check Parameter names}
\usage{
chk_pars(x, x_name = NULL)
}
\arguments{
\item{x}{The object.}

\item{x_name}{A string of the name of object x or NULL.}
}
\value{
\code{NULL}, invisibly. Called for the side effect of throwing an error
if the condition is not met.
}
\description{
Checks if parameter names using \verb{[vld_pars](x)}.
}
\examples{
x <- c("x", "a1._", "X")
chk_pars(x)
y <- c("x[1]", "a1", "a1", "._0")
try(chk_pars(y))
}
