% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete-terms.R
\name{complete_terms.term}
\alias{complete_terms.term}
\title{Complete Terms}
\usage{
\method{complete_terms}{term}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
The modified object.
}
\description{
Adds any absent elements to a terms vector.
}
\details{
The vector is repaired before being completed.
Missing values are ignored.
}
\examples{
complete_terms(as.term(c("b[3]", "b[1]", "b[2]")))
complete_terms(as.term(c("z[2,2]", "z[1,1]")))
}
\seealso{
\code{\link[=term-vector]{term-vector()}}, \code{\link[=repair_terms]{repair_terms()}} and \code{\link[=is.incomplete_terms]{is.incomplete_terms()}}.
}
