% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consistent-term.R
\name{consistent_term}
\alias{consistent_term}
\title{Consistent Terms}
\usage{
consistent_term(x)
}
\arguments{
\item{x}{The object.}
}
\value{
A logical vector indicating whether the number of dimensions is consistent.
}
\description{
Test whether the number of dimensions of terms in the same parameter are consistent.
}
\examples{
consistent_term(term("alpha[1]", "alpha[3]", "beta[1,1]", "beta[2,1]"))
consistent_term(term("alpha[1]", NA_term_, "beta[1,1]", "beta[2]"))
}
\seealso{
\code{\link[=term-vector]{term-vector()}} and \code{\link[=npdims]{npdims()}}
}
