% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consistent-term.R
\name{consistent_terms}
\alias{consistent_terms}
\title{Consistent Terms}
\usage{
consistent_terms(x)
}
\arguments{
\item{x}{A term vector.}
}
\value{
A logical vector indicating whether the number of dimensions is consistent.
}
\description{
Test whether the number of dimensions of terms in the same parameter are consistent.
}
\examples{
consistent_terms(as.term(c("alpha[1]", "alpha[3]", "beta[1,1]", "beta[2,1]")))
}
\seealso{
\code{\link{term-vector}} and \code{\link{npdims}}
}
