% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{is.term}
\alias{is.incomplete_terms}
\alias{is.inconsistent_terms}
\alias{parameters}
\alias{parameters<-}
\alias{set_parameters}
\alias{tdims}
\title{Deprecated Functions}
\usage{
is.term(x)

is.incomplete_terms(x)

is.inconsistent_terms(x)

parameters(x, ...)

parameters(x) <- value

set_parameters(x, pars)

tdims(x)
}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}

\item{x}{The object.}

\item{value}{A character vector of the new parameter names.}

\item{pars}{A character vector of parameter names.}
}
\description{
Deprecated Functions
}
\section{Functions}{
\itemize{
\item \code{is.term()}: Is Term

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}

Replace by \code{\link[=is_term]{is_term()}}

\item \code{is.incomplete_terms()}: Is Incomplete Terms

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}

Replace by \code{\link[=is_incomplete_terms]{is_incomplete_terms()}}

\item \code{is.inconsistent_terms()}: Is Inconsistent Terms

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}

Replace by \code{\link[=is_inconsistent_terms]{is_inconsistent_terms()}}

\item \code{parameters()}: Get Parameters

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}

Replace by \code{\link[=pars]{pars()}}

\item \code{parameters(x) <- value}: Setting Parameters

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}

Replace by pars<-

\item \code{set_parameters()}: Set Parameters

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}

Replace by \code{\link[=set_pars]{set_pars()}}

\item \code{tdims()}: Term Index

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}

Replace by \code{\link[=tindex]{tindex()}}

}}
\keyword{internal}
