% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectTERM.R
\name{detectTERM}
\alias{detectTERM}
\title{Detecting TE-regulation modules from the constructed multi-layer differential expression network}
\usage{
detectTERM(ppi, DEstat, TEstat, num_seed = 100, minModSize = 10,
  permTimes = 100, modularity_p = 0.05, maxModOvlp = 0.5)
}
\arguments{
\item{ppi}{a matrix/dataframe containing gene-gene interactions in two columns}

\item{DEstat}{a list including two dataframes of statistics and p-values of differential expression for mRNAs and RPFs respectively}

\item{TEstat}{a vector of differential translation significance of genes}

\item{num_seed}{the number of seed genes used for searching modules from multi-layer networks, with default value 100}

\item{minModSize}{the number of minimum genes within a module, with default value 10}

\item{permTimes}{permutation times for calculating modulatrity significance of modules, with default value 100}

\item{modularity_p}{threshold for modulatrity significance, with default value 0.05}

\item{maxModOvlp}{threshold for meetmin index defining the overlapping between modules, with default value 0.5}
}
\value{
a list of results including a dataframe of the multi-layer network with edge weights in each network as well as the average weight and a list of identified TE-regulated modules
}
\description{
Detecting TE-regulation modules from the constructed multi-layer differential expression network
}
\examples{
detectTERM()
detectTERM(ppi, DEstat, TEstat,num_seed=100,minModSize=10,permTimes=100,modularity_p=0.05,maxModOvlp=0.5)
}
