% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dims.R
\name{dims}
\alias{dims}
\title{Dimensions of an Object}
\usage{
dims(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{...}{Unused}
}
\value{
A vector or list of the dimensions.
}
\description{
Gets the dimensions of an object.
}
\details{
Unlike base::dim(), dims returns a scalar corresponding to the length of
a vector and two values corresponding to the number of rows and columns 
of a data frame.
}
\examples{
dims(character(0))
dims(1:3)
dims(data.frame())
dims(data.frame(x = 1:3))
}
