% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dims.R
\name{dims.term}
\alias{dims.term}
\title{Dimensions of a default object}
\usage{
\method{dims}{term}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\description{
Dimensions of a default object
}
\examples{
dims(as.term("beta[1,1]"))
dims(as.term(c("beta[1,1]", "beta[1,2]")))
}
\seealso{
\code{\link[universals]{dims}()}
}
