% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dims.R
\name{dims.term_rcrd}
\alias{dims.term_rcrd}
\title{Dimensions}
\usage{
\method{dims}{term_rcrd}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
An integer vector of the dimensions.
}
\description{
Gets the dimensions of an object.
}
\details{
Unlike \code{base::dim()}, dims returns the length of an atomic vector.
}
\examples{

dims(term_rcrd("beta[1,1]"))
dims(term_rcrd("beta[1,1]", "beta[1,2]"))
}
\seealso{
\code{\link[base:dim]{base::dim()}}

Other dimensions: 
\code{\link[universals]{ndims}()},
\code{\link[universals]{npdims}()},
\code{\link[universals]{pdims}()}
}
