% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-inconsistent-terms.R
\name{is.inconsistent_terms}
\alias{is.inconsistent_terms}
\title{Is Inconsistent Terms}
\usage{
is.inconsistent_terms(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Unused.}
}
\value{
A logical scalar indicating whether the object's terms are inconsistent.
}
\description{
Tests whether a term vector has inconsistent elements.
}
\examples{
is.inconsistent_terms(as.term("b[2]"))
is.inconsistent_terms(as.term(c("b[2]", "b[1]")))
is.inconsistent_terms(as.term(c("b[2]", "b[1,1]")))
}
\seealso{
\code{\link[=term-vector]{term-vector()}} and \code{\link[=consistent_term]{consistent_term()}}
}
