% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.term}
\alias{is.term}
\title{Is Term}
\usage{
is.term(x)
}
\arguments{
\item{x}{The object to test.}
}
\value{
A flag indicating whether the test was positive.
}
\description{
Tests whether an R object inherits from class \code{term}.
}
\details{
It does not test the validity of the term elements.
}
\examples{
is.term(c("parameter[2]", "parameter[10]"))
is.term(as.term(c("parameter[2]", "parameter[10]")))
}
\seealso{
\code{\link{term-vector}} and \code{\link{valid_term}()}
}
