% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-inconsistent-terms.R
\name{is_inconsistent_terms}
\alias{is_inconsistent_terms}
\title{Is Inconsistent Terms}
\usage{
is_inconsistent_terms(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
A logical scalar indicating whether the object's terms are inconsistent.
}
\description{
Tests whether a term vector has inconsistent elements.
Returns TRUE if includes missing or invalid terms.
}
\examples{
is_inconsistent_terms(term("b[2]"))
is_inconsistent_terms(term("b[2]", "b[1]"))
is_inconsistent_terms(term("b[2]", "b[1,1]"))
}
\seealso{
\code{\link[=term-vector]{term-vector()}} and \code{\link[=consistent_term]{consistent_term()}}
}
