% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_term}
\alias{is_term}
\title{Is Term}
\usage{
is_term(x)
}
\arguments{
\item{x}{The object.}
}
\value{
A flag indicating whether the test was positive.
}
\description{
Tests whether an R object inherits from S3 class \code{term}.
}
\details{
It does not test the validity of consistency of the term elements.
}
\examples{
is_term(c("parameter[2]", "parameter[10]"))
is_term(term("parameter[2]", "parameter[10]"))
}
\seealso{
\code{\link[=term-vector]{term-vector()}}, \code{\link[=vld_term]{vld_term()}}, \code{\link[=valid_term]{valid_term()}}
and \code{\link[=consistent_term]{consistent_term()}}
}
