% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_term_rcrd2}
\alias{is_term_rcrd2}
\title{Is Term Record}
\usage{
is_term_rcrd2(x)
}
\arguments{
\item{x}{The object.}
}
\value{
A flag indicating whether the test was positive.
}
\description{
Tests whether an R object inherits from S3 class \code{term_rcrd2}.
}
\details{
It does not test the validity of consistency of the term elements.
}
\examples{
is_term_rcrd2(new_term_rcrd2())
}
\seealso{
\code{\link[=valid_term]{valid_term()}}
and \code{\link[=consistent_term]{consistent_term()}}
}
