% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-term.R
\name{NA_term_}
\alias{NA_term_}
\title{Missing Term}
\usage{
NA_term_
}
\description{
A missing term element.
}
\examples{
is_term(NA_term_)
is.na(NA_term_)
}
\seealso{
\code{\link[=term-vector]{term-vector()}}
}
