% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-term-rcrd.R
\name{NA_term_rcrd_}
\alias{NA_term_rcrd_}
\title{Missing Term}
\usage{
NA_term_rcrd_
}
\description{
A missing term element of \code{term_rcrd} type.
}
\examples{
is_term_rcrd(NA_term_)
is.na(NA_term_)
}
\seealso{
\code{\link[=term-vector]{term-vector()}}
}
