% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndims.R
\name{ndims}
\alias{ndims}
\title{Number of Dimensions of an Object}
\usage{
ndims(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{...}{Unused}
}
\value{
A vector or list of the number of dimensions.
}
\description{
Gets the number of dimensions of an object as defined by \code{\link{dims}()}.
}
\examples{
ndims(character(0))
ndims(1:3)
ndims(data.frame())
ndims(data.frame(x = 1:3))
}
