% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-term.R
\name{new_term}
\alias{new_term}
\title{Construct a new term object}
\usage{
new_term(x = character())
}
\arguments{
\item{x}{A character vector.}
}
\description{
Use this function to quickly construct a \link{term} object
from a character vector, without checking the input.
Use \code{\link[=as_term]{as_term()}} to validate the input.
}
\examples{
new_term()
new_term(c("a", "b[1]", "b[2]"))
}
