% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-term-rcrd.R
\name{new_term_rcrd}
\alias{new_term_rcrd}
\title{Construct a New Term Record Object}
\usage{
new_term_rcrd(
  x = data.frame(par = character(), dim = I(list()), stringsAsFactors = FALSE)
)
}
\arguments{
\item{x}{A data frame with columns par and dim.}
}
\description{
Use this function to quickly construct a term_rcrd object.
}
\examples{
new_term_rcrd()
\dontrun{
new_term_rcrd(data.frame(
  par = c("x", "x", "y"), dim = I(list(1, 2, c(2, 2))),
  stringsAsFactors = FALSE
))
}
}
\seealso{
\code{\link[=new_term]{new_term()}}
}
