% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-term-rcrd2.R
\name{new_term_rcrd2}
\alias{new_term_rcrd2}
\title{Construct a new term_rcrd2 object}
\usage{
new_term_rcrd2(
  x = data.frame(par = character(), dim = I(list()), stringsAsFactors = FALSE)
)
}
\arguments{
\item{x}{A data frame with columns par and dim.}
}
\description{
Use this function to quickly construct a term_rcrd2 object.
}
\examples{
new_term_rcrd2()
new_term_rcrd2(data.frame(par = c("x", "x", "y"), dim = I(list(1, 2, c(2,2))),
stringsAsFactors = FALSE))
}
