% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize-terms.R
\name{normalize_terms}
\alias{normalize_terms}
\title{Normalize Terms}
\usage{
normalize_terms(x)
}
\arguments{
\item{x}{The object.}
}
\value{
The normalized term vector.
}
\description{
Normalizes a term vector.
}
\details{
If a parameter such as \code{b}
is a scalar then \code{b[1]} is replaced by \code{b}
but if higher indices are included such as \code{b[2]} then \code{b}
is replaced by \code{b[1]}.
}
\examples{
normalize_terms(new_term(c("b", "b[3]")))
normalize_terms(new_term(c("b[1]", "a[3]")))
}
\seealso{
\code{\link[=term-vector]{term-vector()}} and \code{\link[=repair_terms]{repair_terms()}}
}
