% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/npars.R
\name{npdims}
\alias{npdims}
\alias{npars}
\title{Number of Parameters}
\usage{
npdims(x)

npars(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Unused.}
}
\value{
A count of the number of parameters.
}
\description{
Gets the number of parameters for an object.
}
\examples{
npars(as.term(c("alpha[1]", "alpha[2]", "beta[1,1]", "beta[2,1]",
"beta[1,2]", "beta[2,2]", "sigma")))
}
