% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npars-scalar.R
\name{npars_scalar}
\alias{npars_scalar}
\alias{npars_scalar.default}
\title{Number of Scalar Parameters}
\usage{
npars_scalar(x, ...)

\method{npars_scalar}{default}(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Unused.}
}
\value{
An integer scalar of the number of scalar parameters.
}
\description{
Gets the number of parameters of an object as returned by \code{\link[=pars_scalar]{pars_scalar()}}.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Number of Scalar Parameters
}}

\seealso{
Other {parameters}: 
\code{\link{pars.term}()},
\code{\link{pars_scalar}()},
\code{\link{pars_terms}()}
}
\concept{{parameters}}
