% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npdims.R
\name{npdims.term}
\alias{npdims.term}
\title{Number of Dimensions of Each Parameter}
\usage{
\method{npdims}{term}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
A named integer vector of the number of dimensions of each parameter.
}
\description{
Gets the number of the dimensions of each parameter of an object.

The default methods returns the length of each element of \code{\link[universals:pdims]{pdims()}}
as an integer vector.
}
\examples{
npdims(term("alpha[1]", "alpha[3]", "beta[1,1]", "beta[2,1]"))
}
\seealso{
Other dimensions: 
\code{\link[universals]{dims}()},
\code{\link[universals]{ndims}()},
\code{\link[universals]{pdims}()}
}
