% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\alias{arguments}
\alias{args}
\title{Parameter Descriptions for term Package}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}

\item{x}{The object.}

\item{name}{A string specifying the name of the parameter.}

\item{validate}{A string specifying the level of the validation.
The possible values in order of increasing strictness
are 'class', 'valid', 'consistent' and 'complete'.}

\item{repair}{A flag specifying whether to repair terms.}

\item{normalize}{A flag specifying whether to normalize terms.}

\item{select}{A character vector of the names of the parameters to include in the subsetted object.}

\item{value}{A character vector of the new parameter names.}

\item{pars}{A character vector of parameter names.}

\item{scalar}{A flag specifying whether to by default return all parameters (NULL), or only scalar parameters (TRUE) or only non-scalar parameters (FALSE).}

\item{terms}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}} A flag specifying whether to
return the parameter name for each term element. Use \code{\link[=pars_terms]{pars_terms()}} instead.}

\item{incomparables}{Ignored.}
}
\description{
Default parameter descriptions which may be overridden in individual
functions.
}
\details{
A flag is a non-missing logical scalar.

A string is a non-missing character scalar.
}
\keyword{internal}
