% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pars.R
\name{pars.default}
\alias{pars.default}
\title{Parameter Names}
\usage{
\method{pars}{default}(x, scalar = FALSE, ...)
}
\arguments{
\item{x}{An object.}

\item{scalar}{A flag specifying whether to only include scalar parameters.}

\item{...}{Other arguments passed to methods.}
}
\value{
A character vector of the names of the parameters.
}
\description{
Gets the parameter names.
}
\examples{
pars.foobar <- function(x, ...) {
  NotYetImplemented()
  # replace with code to get pars for an object of class 'foobar'
}
}
\seealso{
Other parameters: 
\code{\link[universals]{npars}()},
\code{\link[universals]{set_pars}()}
}
