% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/pars.R
\name{parameters}
\alias{parameters}
\alias{pararameters<-}
\alias{set_parameters}
\alias{pars}
\alias{pars<-}
\alias{set_pars}
\alias{pars.default}
\alias{pars.character}
\alias{pars.term}
\title{Parameter Names}
\usage{
parameters(x, ...)

pararameters(x) <- value

set_parameters(x, pars)

pars(x, ...)

pars(x) <- value

set_pars(x, pars)

\method{pars}{default}(x, scalar_only = FALSE, terms = FALSE, ...)

\method{pars}{character}(x, scalar_only = FALSE, terms = FALSE, ...)

\method{pars}{term}(x, scalar_only = FALSE, terms = FALSE, ...)
}
\arguments{
\item{x}{An R object.}

\item{...}{Unused.}

\item{value}{A character vector of the new parameter names.}

\item{pars}{A character vector of the new parameter names.}

\item{scalar_only}{A flag specifying whether to only get the names of
parameters with one term.}

\item{terms}{A flag specifying whether to return the parameter name
for each term element.}
}
\value{
A character vector of the parameter names.
}
\description{
Gets or sets the parameter names for an object.
It assumes that the terms do not require repairing.
}
\details{
\code{parameters()} is deprecated.

\code{parameters<-} is deprecated.

\code{set_parameters()} is deprecated.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Parameter names for a default object

\item \code{character}: Parameter names for a character vector

\item \code{term}: Parameter names for a term vector
}}

\examples{
term <- as.term(c("alpha[1]", "alpha[2]", "beta[1,1]", "beta[2,1]",
"beta[1,2]", "beta[2,2]", "sigma"))
pars(term)
pars(term, scalar_only = TRUE)
pars(term, terms = TRUE)
}
\seealso{
\code{\link{repair_terms}()}
}
