% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pars.R
\name{pars.term}
\alias{pars.term}
\title{Parameter Names}
\usage{
\method{pars}{term}(x, scalar = NULL, terms = FALSE, ...)
}
\arguments{
\item{x}{An object.}

\item{scalar}{A flag specifying whether to by default return all parameters (NULL), or only scalar parameters (TRUE) or only non-scalar parameters (FALSE).}

\item{terms}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}} A flag specifying whether to
return the parameter name for each term element. Use \code{\link[=pars_terms]{pars_terms()}} instead.}

\item{...}{Other arguments passed to methods.}
}
\value{
A character vector of the names of the parameters.
}
\description{
Gets the parameter names.
}
\details{
The \code{terms} argument is \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}.
Use \code{\link[=pars_terms]{pars_terms()}} to get the parameter name for each term element.
}
\examples{
term <- term(
  "alpha[1]", "alpha[2]", "beta[1,1]", "beta[2,1]",
  "beta[1,2]", "beta[2,2]", "sigma", NA
)
pars(term)
pars(term, scalar = TRUE)
pars(term, scalar = FALSE)
}
\seealso{
\link[universals:pars]{universals::pars}

Other parameters:
\code{\link[=pars.character]{pars.character()}},
\code{\link[=pars.default]{pars.default()}},
\code{\link[=pars.term_rcrd]{pars.term_rcrd()}},
\code{\link[=pars_terms]{pars_terms()}}
}
\concept{parameters}
