% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pars.R
\name{pars.term_rcrd}
\alias{pars.term_rcrd}
\title{Parameter Names}
\usage{
\method{pars}{term_rcrd}(x, scalar = NULL, ...)
}
\arguments{
\item{x}{An object.}

\item{scalar}{A flag specifying whether to by default return all parameters (NULL), or only scalar parameters (TRUE) or only non-scalar parameters (FALSE).}

\item{...}{Other arguments passed to methods.}
}
\value{
A character vector of the names of the parameters.
}
\description{
Gets the parameter names.
}
\examples{
term <- term(
  "alpha[1]", "alpha[2]", "beta[1,1]", "beta[2,1]",
  "beta[1,2]", "beta[2,2]", "sigma", NA
)
pars(term)
pars(term, scalar = TRUE)
pars(term, scalar = FALSE)
}
\seealso{
\link[universals:pars]{universals::pars}

Other parameters: 
\code{\link{pars.character}()},
\code{\link{pars.default}()},
\code{\link{pars.term}()},
\code{\link{pars_terms}()}
}
\concept{parameters}
