% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pars-scalar.R
\name{pars_scalar.default}
\alias{pars_scalar.default}
\title{Number of Scalar Parameters}
\usage{
\method{pars_scalar}{default}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
An integer scalar of the number of scalar parameters.
}
\description{
Gets the number of scalar parameters of an object.
}
\seealso{
\code{\link[universals:pars_scalar]{pars_scalar()}}

Other {MCMC dimensions}: 
\code{\link[universals]{nchains}()},
\code{\link[universals]{niters}()},
\code{\link[universals]{npars}()},
\code{\link[universals]{nsams}()},
\code{\link[universals]{nsims}()},
\code{\link[universals]{nterms}()}

Other {parameters}: 
\code{\link[universals]{npars}()},
\code{\link[universals]{pars_scalar}()},
\code{\link[universals]{pars}()},
\code{\link[universals]{set_pars}()}
}
