% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pars-scalar.R
\name{pars_scalar}
\alias{pars_scalar}
\alias{pars_scalar.term}
\title{Scalar Parameters}
\usage{
pars_scalar(x, ...)

\method{pars_scalar}{term}(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Unused.}
}
\value{
A character vector of the parameter names.
}
\description{
Gets the name of each scalar parameter.
}
\section{Methods (by class)}{
\itemize{
\item \code{term}: Scalar Parameter Names for term vector
}}

\examples{
term <- as.term(c(
  "alpha[1]", "alpha[2]", "beta[1,1]", "beta[2,1]",
  "beta[1,2]", "beta[2,2]", "sigma", NA
))
pars_scalar(term)
}
\concept{{parameters}}
