% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pars-scalar.R
\name{pars_scalar.term}
\alias{pars_scalar.term}
\title{Number of Scalar Parameters}
\usage{
\method{pars_scalar}{term}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
An integer scalar of the number of scalar parameters.
}
\description{
Gets the number of scalar parameters of an object.
}
\examples{
term <- term(
  "alpha[1]", "alpha[2]", "beta[1,1]", "beta[2,1]",
  "beta[1,2]", "beta[2,2]", "sigma", NA
)
pars_scalar(term)
}
\seealso{
\code{\link[universals:pars_scalar]{pars_scalar()}}

Other {MCMC dimensions}: 
\code{\link[universals]{nchains}()},
\code{\link[universals]{niters}()},
\code{\link[universals]{npars}()},
\code{\link[universals]{nsams}()},
\code{\link[universals]{nsims}()},
\code{\link[universals]{nterms}()}

Other {parameters}: 
\code{\link[universals]{npars}()},
\code{\link[universals]{pars_scalar}()},
\code{\link[universals]{pars}()},
\code{\link[universals]{set_pars}()}
}
