% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pars-terms.R
\name{pars_terms}
\alias{pars_terms}
\title{Term Parameters}
\usage{
pars_terms(x, scalar = NA, ...)
}
\arguments{
\item{x}{A term vector.}

\item{scalar}{A logical scalar specifying whether to include
all parameters (NA), only scalars (TRUE) or all parameters
except scalars (FALSE).}

\item{...}{Unused.}
}
\value{
A character vector of the term parameter names.
}
\description{
Gets the name of each parameter for each term.
}
\examples{
term <- as.term(c(
  "alpha[1]", "alpha[2]", "beta[1,1]", "beta[2,1]",
  "beta[1,2]", "beta[2,2]", "sigma", NA
))
pars_terms(term)
}
