% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pars-terms.R
\name{pars_terms.term}
\alias{pars_terms.term}
\title{Term Parameters}
\usage{
\method{pars_terms}{term}(x, scalar = NA, ...)
}
\arguments{
\item{x}{An object.}

\item{scalar}{A logical scalar specifying whether to include
all parameters (NA), only scalars (TRUE) or all parameters
except scalars (FALSE).}

\item{...}{Other arguments passed to methods.}
}
\value{
A character vector of the term parameter names.
}
\description{
Gets the name of each parameter for each term.
}
\examples{
term <- as.term(c(
  "alpha[1]", "alpha[2]", "beta[1,1]", "beta[2,1]",
  "beta[1,2]", "beta[2,2]", "sigma", NA
))
pars_terms(term)
}
\seealso{
Other {parameters}: 
\code{\link[universals]{npars}()},
\code{\link[universals]{pars}()},
\code{\link[universals]{set_pars}()}

Other {terms}: 
\code{\link[universals]{complete_terms}()},
\code{\link[universals]{nterms}()}
}
