% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdims.R
\name{pdims}
\alias{pdims}
\alias{pdims.term}
\title{Parameter Dimensions}
\usage{
pdims(x, ...)

\method{pdims}{term}(x, ...)
}
\arguments{
\item{x}{The object}

\item{...}{Unused.}
}
\description{
Gets the parameter dimensions of an object.
}
\section{Methods (by class)}{
\itemize{
\item \code{term}: Parameter dimensions for a term vector
}}

\examples{
pdims(as.term(c("alpha[1]", "alpha[2]", "beta[1,1]", "beta[2,1]",
"beta[1,2]", "beta[2,2]", "sigma")))
}
