% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_TERM.R
\name{plot_TERM}
\alias{plot_TERM}
\title{Plot a module identified by TERM.}
\usage{
plot_TERM(TEmodule, layout_style = "layout.fruchterman.reingold",
  gene2symbol = NULL)
}
\arguments{
\item{TEmodule}{a module identified by TERM.}

\item{layout_style}{the layout style of the graph for the module, with the default value "layout.fruchterman.reingold".}

\item{gene2symbol}{a file converting the original gene identifiers to alternative identifiers, with the default value NULL.}
}
\value{
visualization for the network of the input module 
where the core and border colors of the nodes denoting differential expression of mRNA and RPF 
and the square shape of the nodes indicating the significantly translated genes.
}
\description{
Plot a module identified by TERM.
}
\examples{
plot_TERM()
plot_TERM(TEmodule, layout_style="layout.fruchterman.reingold", gene2symbol=NULL  )
}
