% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repair-terms.R
\name{repair_terms}
\alias{repair_terms}
\title{Repair Terms}
\usage{
repair_terms(x)
}
\arguments{
\item{x}{A term vector.}
}
\value{
The repaired term vector.
}
\description{
Repairs values in an terms vector.
}
\details{
Initially removes spaces.
Empty strings (\code{""}) are dropped.
If a parameter such as \code{b} is a scalar then \code{b[1]} is replaced by \code{b}
but if higher indices are included such as \code{b[2]} then \code{b}
is replaced by \code{b[1]}.
}
\examples{
complete_terms(term("b[1]"))
complete_terms(term("b[3]"), as.term(c("b[3]", "b[1]", "b[2]")))
complete_terms(term(c("z[2,2]", "z[1,1]")))
}
