% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-pars.R
\name{set_pars.term_rcrd}
\alias{set_pars.term_rcrd}
\title{Set Parameter Names}
\usage{
\method{set_pars}{term_rcrd}(x, value, ...)
}
\arguments{
\item{x}{An object.}

\item{value}{A character vector of the new parameter names.}

\item{...}{Other arguments passed to methods.}
}
\value{
The modified object.
}
\description{
Sets an object's parameter names.

The assignment version \verb{pars<-()} forwards to \code{set_pars()}.
}
\details{
\code{value} must be a unique character vector of the same length as the
object's parameters.
}
\examples{
term_rcrd <- term_rcrd("b[2]", "a[1]", "b[3,3]")
set_pars(term_rcrd, c("x", "y"))
}
\seealso{
Other parameters: 
\code{\link[universals]{npars}()},
\code{\link[universals]{pars}()}
}
