% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset.term_rcrd}
\alias{subset.term_rcrd}
\title{Subset Term Record}
\usage{
\method{subset}{term_rcrd}(x, pars = NULL, ...)
}
\arguments{
\item{x}{The object.}

\item{pars}{A character vector of parameter names.}

\item{...}{Unused.}
}
\value{
The modified term vector.
}
\description{
Subsets a term_rcrd.
}
\examples{
term_rcrd <- term_rcrd(
  "alpha[1]", "alpha[2]", "beta[1,1]", "beta[2,1]",
  "beta[1,2]", "beta[2,2]", "sigma"
)
\dontrun{
subset(term_rcrd, "beta")
subset(term_rcrd, c("alpha", "sigma"))
}
}
\seealso{
\code{\link[=term_rcrd_object]{term_rcrd_object()}}
}
