% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term-data-frame.R
\name{term-data-frame}
\alias{term-data-frame}
\title{Term Data Frame}
\description{
A \code{term} data frame is a tibble with a term column and one or more
numeric columns.
}
\details{
By default the term column is named term and is the first column
and the first numeric column is name n1 and is the second column.
By default subsequent numeric columns are named n1, n2 etc.
}
\examples{
term_df(matrix(1:4, 2))
}
