% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term-object.R
\name{term-object}
\alias{term-object}
\alias{term_object}
\alias{term-vector}
\alias{term_vector}
\title{Term Object}
\description{
A \code{term} object is a S3 vector of parameter terms.
}
\examples{
as.term(c("a[1]", "a[2]"))
term(matrix(1:4, 2))
str(as.term("a"))
}
