% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term-package.R
\docType{package}
\name{term-package}
\alias{term-package}
\title{term: Create, Manipulate and Query Parameter Terms}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Creates, manipulates, queries and repairs vectors of parameter terms. Parameter terms are the labels used to reference values in vectors, matrices and arrays. They represent the names in coefficient tables and the column names in 'mcmc' and 'mcmc.list' objects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://poissonconsulting.github.io/term/}
  \item \url{https://github.com/poissonconsulting/term}
  \item Report bugs at \url{https://github.com/poissonconsulting/term/issues}
}

}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID})

Authors:
\itemize{
  \item Kirill Müller (\href{https://orcid.org/0000-0002-1416-3412}{ORCID})
}

Other contributors:
\itemize{
  \item Ayla Pearson (\href{https://orcid.org/0000-0001-7388-1222}{ORCID}) [contributor]
  \item Evan Amies-Galonski (\href{https://orcid.org/0000-0003-1096-2089}{ORCID}) [contributor]
  \item Poisson Consulting [copyright holder, funder]
}

}
\keyword{internal}
