% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term.R
\name{term}
\alias{term}
\title{The main function performs the whole pipeline of TERM to identify TE-regulated modules.}
\usage{
term(raw_rna, raw_ribo, rna_label, ribo_label, baseLevel, raw_ppi,
  minCounts = 10, minCountsProportion = 1, num_seed = 100,
  minModSize = 10, permTimes = 100, modularity_p = 0.05,
  maxModOvlp = 0.5)
}
\arguments{
\item{raw_rna}{raw RNA-seq read counts matrix}

\item{raw_ribo}{raw Ribo-seq read counts matrix}

\item{rna_label}{the experiment conditions corresponding to the samples in raw_rna}

\item{ribo_label}{the experiment conditions corresponding to the samples in raw_ribo}

\item{baseLevel}{the condition used as the base level for comparisons of differential expression and differential translation}

\item{raw_ppi}{a matrix/dataframe containing gene-gene interactions in two columns}

\item{minCounts}{the number of minimum read count of a gene in a sample, with default value 10}

\item{minCountsProportion}{the proportion of samples where a gene exhibits read counts larger than the minCounts in a same condition, with default value 1}

\item{num_seed}{the number of seed genes used for searching modules from multi-layer networks, with default value 100}

\item{minModSize}{the number of minimum genes within a module, with default value 10}

\item{permTimes}{permutation times for calculating modulatrity significance of modules, with default value 100}

\item{modularity_p}{threshold for modulatrity significance, with default value 0.05}

\item{maxModOvlp}{threshold for meetmin index defining the overlapping between modules, with default value 0.5}
}
\value{
a list of results including a dataframe of mRNA and RPF differential expression and differential translation of genes, 
a dataframe of the multi-layer network with edge weights in each network as well as the average weight, 
a list of identified TE-regulated modules
}
\description{
The main function performs the whole pipeline of TERM to identify TE-regulated modules.
}
\examples{
term()
term(raw_rna, raw_ribo, rna_label, ribo_label, baseLevel, raw_ppi, minCounts=10, minCountsProportion=1,num_seed=100,minModSize=10,permTimes=100, modularity_p=0.05,maxModOvlp=0.5 )
}
