% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{term_deprecated}
\alias{term_deprecated}
\alias{parameters}
\alias{parameters<-}
\alias{pararameters<-}
\alias{set_parameters}
\alias{tdims}
\alias{is.term}
\title{Deprecated Functions}
\usage{
parameters(x, ...)

parameters(x) <- value

pararameters(x) <- value

set_parameters(x, pars)

tdims(x)

is.term(x)
}
\arguments{
\item{x}{The object.}

\item{...}{Unused.}

\item{value}{A character vector of the new parameter names.}

\item{pars}{A character vector of the new parameter names.}
}
\description{
Deprecated \code{term()} functions
}
\section{Functions}{
\itemize{
\item \code{parameters}: Get Parameters

\lifecycle{soft-deprecated}

Replace by \code{\link[=pars]{pars()}}

\item \code{parameters<-}: Setting Parameters

\lifecycle{soft-deprecated}

Replace by pars<-

\item \code{pararameters<-}: Setting Parameters

\lifecycle{soft-deprecated}

Replace by pars<-

\item \code{set_parameters}: Set Parameters

\lifecycle{soft-deprecated}

Replace by \code{\link[=set_pars]{set_pars()}}

\item \code{tdims}: Term Index

\lifecycle{soft-deprecated}

Replace by \code{\link[=tindex]{tindex()}}

\item \code{is.term}: Is Term

\lifecycle{soft-deprecated}

Replace by \code{\link[=is_term]{is_term()}}
}}

\keyword{internal}
