% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term-df.R
\name{term_df}
\alias{term_df}
\alias{term_df.numeric}
\title{Term Data Frame}
\usage{
term_df(x, ...)

\method{term_df}{numeric}(x, name = "par", ...)
}
\arguments{
\item{x}{An R object.}

\item{...}{Unused.}

\item{name}{A string of the name for the parameter.}
}
\value{
A term data frame.
}
\description{
Creates a term data frame from object.
}
\section{Methods (by class)}{
\itemize{
\item \code{numeric}: Term data frame for a numeric vector
}}

\examples{
term_df(matrix(1:4, 2))
}
\seealso{
\code{\link{term-vector}}
}
