% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term-rcrd.R
\name{term_rcrd}
\alias{term_rcrd}
\alias{term_rcrd-object}
\alias{term_rcrd_object}
\title{Term Record}
\usage{
term_rcrd(...)
}
\arguments{
\item{...}{Unnamed values are term values,
named values describe the parameter in the name
and the dimensionality in the value.}
}
\value{
A term_rcrd vector.
}
\description{
Creates a term_rcrd from values.
This function checks that all terms are valid
but does not require stronger levels of consistency,
see \code{chk_valid()} for details.
}
\examples{
term_rcrd()
\dontrun{
term_rcrd("p", "q[1]", "q[2]", "q[3]")
term_rcrd("q[1]", "q[2]", "q[3]")
}
}
\seealso{
\code{\link[=dims]{dims()}} and \code{\link[=pdims]{pdims()}}
}
