% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term-rcrd2.R
\name{term_rcrd2}
\alias{term_rcrd2}
\alias{term_rcrd2-object}
\alias{term_rcrd2_object}
\title{Term Record}
\usage{
term_rcrd2(...)
}
\arguments{
\item{...}{Unnamed values are term values,
named values describe the parameter in the name
and the dimensionality in the value.}
}
\value{
A term_rcrd2 vector.
}
\description{
Creates a term_rcrd2 from values.
This function checks that all terms are valid
but does not require stronger levels of consistency,
see \code{chk_valid()} for details.
}
\examples{
term_rcrd2()
term_rcrd2("p", "q[1]", "q[2]", "q[3]")
term_rcrd2("q[1]", "q[2]", "q[3]")
}
\seealso{
\code{\link[=dims]{dims()}} and \code{\link[=pdims]{pdims()}}
}
