% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/tindex.R
\name{tdims}
\alias{tdims}
\alias{tindex}
\title{Term Index}
\usage{
tdims(x)

tindex(x, ...)
}
\arguments{
\item{x}{The object}

\item{...}{Unused.}
}
\description{
Gets the index for each term of an object.
}
\details{
\code{tdims()} is deprecated.
}
\examples{
tindex(c("alpha[1]", "alpha[2]", "beta[1,1]", "beta[2,1]"))
}
