% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique.R
\name{unique.term}
\alias{unique.term}
\title{Extract Unique Terms}
\usage{
\method{unique}{term}(x, incomparables = FALSE, repair_terms = TRUE,
  ...)
}
\arguments{
\item{x}{A term vector.}

\item{incomparables}{Ignored.}

\item{repair_terms}{A flag specifying whether the vector should be repaired
before extracting the unique values.}

\item{...}{Unused.}
}
\value{
A vector of unique terms.
}
\description{
Extracts unique values from a term vector. 
By default the vector is repaired before extracting the unique values.
}
\examples{
unique(term(c("a", "a[2 ]", "a [1]")))
unique(term(c("a", "b[2 ]", "a [1]")))
}
\seealso{
\code{\link{repair_terms}()}
}
