% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec-cast.R
\name{vec_cast.term}
\alias{vec_cast.term}
\title{Cast a term vector to specified type}
\usage{
\method{vec_cast}{term}(x, to, ...)
}
\arguments{
\item{x}{Vectors to cast.}

\item{to}{Type to cast to. If \code{NULL}, \code{x} will be returned as is.}

\item{...}{For \code{vec_cast_common()}, vectors to cast. For
\code{vec_cast()} and \code{vec_restore()}, these dots are only for future
extensions and should be empty.}
}
\description{
Cast a term vector to specified type
}
\examples{
vec_cast(new_term(c("a[1]", "a[2]")), "")
vec_cast(c("a[1]", "a[2]"), new_term(""))
}
\seealso{
\code{\link[vctrs:vec_cast]{vctrs::vec_cast()}}
}
