% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec-ptype2.R
\name{vec_ptype2.term}
\alias{vec_ptype2.term}
\title{Find the common type for a pair of vector types}
\usage{
\method{vec_ptype2}{term}(x, y, ...)
}
\arguments{
\item{x}{Vector types.}

\item{y}{Vector types.}

\item{...}{These dots are for future extensions and must be empty.}
}
\description{
\code{vec_ptype2()} finds the common type for a pair of vectors, or dies trying.
It forms the foundation of the vctrs type system, along with \code{\link[vctrs:vec_cast]{vec_cast()}}.
This powers type coercion but should not usually be called directly;
instead call \code{\link[vctrs:vec_ptype_common]{vec_ptype_common()}}.
}
\seealso{
\code{\link[vctrs:vec_ptype2]{vctrs::vec_ptype2()}}
}
