% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vld.R
\name{vld_pars}
\alias{vld_pars}
\title{Validate Parameter Names}
\usage{
vld_pars(x)
}
\arguments{
\item{x}{The object.}
}
\value{
A flag indicating whether the condition was met.
}
\description{
Validates a character vector of parameter names.
}
\details{
The character vector must consist of unique, non-missing values.
}
\examples{
vld_pars(c("x", "a1._", "X"))
vld_pars(c("x[1]", "a1", "a1", "._0"))
}
\seealso{
\code{\link[=vld_pars]{vld_pars()}}
}
