% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-character.R, R/as-term.R
\name{as.character.term}
\alias{as.character.term}
\alias{as.term}
\alias{as.term.character}
\alias{as.term.default}
\title{Coerce to a term Vector}
\usage{
\method{as.character}{term}(x, ...)

as.term(x, ...)

\method{as.term}{character}(x, ...)

\method{as.term}{default}(x, name = "parameter", ...)
}
\arguments{
\item{x}{The object to coerce.}

\item{...}{Unused.}

\item{name}{A string of the name for the parameter.}
}
\description{
Coerces an R object to a term vector.
}
\details{
A term vector is a S3 vector of parameter terms.
}
\section{Methods (by class)}{
\itemize{
\item \code{term}: Coerce term vector to character vector

\item \code{character}: Coerce character vector to term vector

\item \code{default}: Coerce default object to term vector
}}

\examples{
as.term(matrix(1:4, 2))
as.term(c("parm3[10]", "parm3[2]", "parm[2,2]", "parm[1,1]")) 
}
