% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-term.R
\name{as.term}
\alias{as.term}
\title{Coerce to a term Vector}
\usage{
as.term(x, ...)
}
\arguments{
\item{x}{The object to coerce}

\item{...}{Unused.}
}
\description{
Coerces an R object to a term vector.
}
\details{
A term vector is a S3 vector of parameter terms.
}
\examples{
term <- as.term(c("parm3[10]", "parm3[2]", "parm[2,2]", "parm[1,1]")) 
term
sort(term)
str(term)
}
