% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete-terms.R
\name{complete_terms}
\alias{complete_terms}
\alias{complete_terms.default}
\alias{complete_terms.term}
\title{Complete Terms}
\usage{
complete_terms(x, ...)

\method{complete_terms}{default}(x, ...)

\method{complete_terms}{term}(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{...}{Unused}
}
\value{
The repaired and complete term vector.
}
\description{
Completes any missing values to a terms vector.
}
\details{
The vector is repaired before being completed.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Complete terms of a default object

\item \code{term}: Complete terms of a terms vector
}}

\examples{
complete_terms(term("b[1]"))
complete_terms(term("b[3]"), as.term(c("b[3]", "b[1]", "b[2]")))
complete_terms(term(c("z[2,2]", "z[1,1]")))
}
\seealso{
\code{\link{repair_terms}()} and \code{\link{is.incomplete_terms}()}.
}
