% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete-terms.R
\name{complete_terms}
\alias{complete_terms}
\alias{complete_terms.term}
\title{Complete Terms}
\usage{
complete_terms(x, ...)

\method{complete_terms}{term}(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Unused.}
}
\value{
The repaired and complete term vector.
}
\description{
Adds any absent elements to a terms vector.
}
\details{
The vector is repaired before being completed.
Missing values are ignored.
}
\section{Methods (by class)}{
\itemize{
\item \code{term}: Complete terms of a terms vector
}}

\examples{
complete_terms(as.term(c("b[3]", "b[1]", "b[2]")))
complete_terms(as.term(c("z[2,2]", "z[1,1]")))
}
\seealso{
\code{\link[=term-vector]{term-vector()}}, \code{\link[=repair_terms]{repair_terms()}} and \code{\link[=is.incomplete_terms]{is.incomplete_terms()}}.
}
