% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dims.R
\name{dims.term}
\alias{dims.term}
\title{Dimensions of a term}
\usage{
\method{dims}{term}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
An integer vector of the dimensions.
}
\description{
Gets the dimensions of an object.
Unlike \code{base::dim()}, dims works as intuitively expected
for vectors and data.frames.
}
\examples{
dims(as.term("beta[1,1]"))
dims(as.term(c("beta[1,1]", "beta[1,2]")))
}
\seealso{
\code{\link[base:dim]{base::dim()}}
}
