% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-incomplete-terms.R
\name{is.incomplete_terms}
\alias{is.incomplete_terms}
\alias{is.incomplete_terms.term}
\title{Is Incomplete Terms}
\usage{
is.incomplete_terms(x, ...)

\method{is.incomplete_terms}{term}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Unused}
}
\value{
A flag indicating whether the object's terms are incomplete.
}
\description{
Tests whether an objects has missing terms.
}
\details{
The test is performed after the objects terms have been repaired.
}
\section{Methods (by class)}{
\itemize{
\item \code{term}: Test whether a term vectors terms are incomplete
}}

\examples{
is.incomplete_terms(as.term("b[2]"))
is.incomplete_terms(as.term(c("b[2]", "b[1]")))
}
\seealso{
\code{\link{repair_terms}()} and \code{\link{complete_terms}()}
}
