% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-incomplete-terms.R
\name{is.incomplete_terms}
\alias{is.incomplete_terms}
\alias{is.incomplete_terms.term}
\title{Is Incomplete Terms}
\usage{
is.incomplete_terms(x, ...)

\method{is.incomplete_terms}{term}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Unused}
}
\value{
A logical scalar indicating whether the object's terms are incomplete.
}
\description{
Tests whether a term vector has absent elements.
}
\section{Methods (by class)}{
\itemize{
\item \code{term}: Test whether a term vector is incomplete
}}

\examples{
is.incomplete_terms(as.term("b[2]"))
is.incomplete_terms(as.term(c("b[2]", "b[1]")))
}
\seealso{
\code{\link[=term-vector]{term-vector()}} and \code{\link[=complete_terms]{complete_terms()}}
}
