% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.term}
\alias{is.term}
\title{Is Term}
\usage{
is.term(x)
}
\arguments{
\item{x}{The object to test.}
}
\value{
A flag indicating whether the test was positive.
}
\description{
Test whether an R object inherits from class term.
}
\examples{
is.term(c("parameter[2]", "parameter[10]"))
is.term(as.term(c("parameter[2]", "parameter[10]")))
}
