% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.term}
\alias{is.term}
\title{Is Term}
\usage{
is.term(x)
}
\arguments{
\item{x}{The object.}
}
\value{
A flag indicating whether the test was positive.
}
\description{
Tests whether an R object inherits from class \code{term}.
}
\details{
It does not test the validity of consistency of the term elements.
}
\examples{
is.term(c("parameter[2]", "parameter[10]"))
is.term(as.term(c("parameter[2]", "parameter[10]")))
}
\seealso{
\code{\link[=term-vector]{term-vector()}}, \code{\link[=valid_term]{valid_term()}}
and \code{\link[=consistent_term]{consistent_term()}}
}
