% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-incomplete-terms.R
\name{is_incomplete_terms}
\alias{is_incomplete_terms}
\title{Is Incomplete Terms}
\usage{
is_incomplete_terms(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
A logical scalar indicating whether the object's terms are incomplete.
}
\description{
Tests whether a term vector has absent elements.
The vector should not require repairing.
}
\examples{
is_incomplete_terms(term("b[2]"))
is_incomplete_terms(term("b[2]", "b[1]"))
is_incomplete_terms(term("b[2]", "b[1]", "b[1]"))
}
\seealso{
\code{\link[=term-vector]{term-vector()}} and \code{\link[=complete_terms]{complete_terms()}}
}
