% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-term.R
\docType{data}
\name{NA_term_}
\alias{NA_term_}
\title{Missing Term}
\format{
An object of class \code{term} (inherits from \code{vctrs_vctr}) of length 1.
}
\usage{
NA_term_
}
\description{
A missing term element.
}
\examples{
is_term(NA_term_)
is.na(NA_term_)
}
\seealso{
\code{\link[=term-vector]{term-vector()}}
}
\keyword{datasets}
