% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-term-rcrd.R
\docType{data}
\name{NA_term_rcrd_}
\alias{NA_term_rcrd_}
\title{Missing Term}
\format{
An object of class \code{term_rcrd} (inherits from \code{vctrs_rcrd}, \code{vctrs_vctr}) of length 1.
}
\usage{
NA_term_rcrd_
}
\description{
A missing term element of \code{term_rcrd} type.
}
\examples{
is_term_rcrd(NA_term_)
is.na(NA_term_)
}
\seealso{
\code{\link[=term-vector]{term-vector()}}
}
\keyword{datasets}
