% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-term.R
\name{new_term}
\alias{new_term}
\title{Construct a New Term Object}
\usage{
new_term(x = character())
}
\arguments{
\item{x}{A character vector.}
}
\description{
Use this function to quickly construct a \link{term} object
from a character vector, without checking the input.
Use \code{\link[=term]{term()}} to repair the input.
}
\examples{
new_term()
new_term(c("a", "b[1]", "b[2]"))

# Terms are not checked for validity:
new_term("r[")
repair_terms(new_term("r["))
}
\seealso{
\code{\link[=new_term_rcrd]{new_term_rcrd()}}
}
