% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npars.R
\name{npars}
\alias{npars}
\title{Number of Parameters}
\usage{
npars(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Unused.}
}
\value{
A count of the number of parameters.
}
\description{
Gets the number of parameters of an object.
}
\examples{
npars(c("alpha[1]", "alpha[2]", "beta[1,1]", "beta[2,1]"))
}
