% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npars.R
\name{npars}
\alias{npars}
\alias{npars.term}
\title{Number of Parameters}
\usage{
npars(x, ...)

\method{npars}{term}(x, scalar_only = FALSE, repair = TRUE,
  na.rm = FALSE, ...)
}
\arguments{
\item{x}{An R object.}

\item{...}{Unused.}

\item{scalar_only}{A flag specifying whether to only count parameters with one term.}

\item{repair}{A flag specifying whether to repair the term vector.}

\item{na.rm}{A flag specifying whether to remove missing term elements
before getting the number of parameters.}
}
\value{
A count of the number of parameters.
}
\description{
Gets the number of parameters of an object as returned by \code{\link{pars}()}.
}
\details{
By default (\code{na.rm = TRUE}) the function repairs the vector and 
returns \code{NA_integer_} if the term vector includes missing term elements.
}
\section{Methods (by class)}{
\itemize{
\item \code{term}: Number of parameters of term vector
}}

\examples{
npars(as.term(c("sigma", "alpha[1]", "alpha[2]", "beta[1,1]", "beta[2,1]")))
npars(as.term(c("sigma", "alpha[1]", "alpha[2]")), scalar_only = TRUE)
}
\seealso{
\code{\link{pars}()}, \code{\link{term-vector}} 
and \code{\link{repair_terms}()}
}
