% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npars.R
\name{npars.term}
\alias{npars.term}
\title{Number of Parameters of a term}
\usage{
\method{npars}{term}(x, scalar = NA, ...)
}
\arguments{
\item{x}{An object.}

\item{scalar}{A flag specifying whether to only count parameters with one term.}

\item{...}{Other arguments passed to methods.}
}
\value{
An integer vector of the dimensions.
}
\description{
Gets the dimensions of an object.
Unlike \code{base::dim()}, dims works as intuitively expected
for vectors and data.frames.
}
\examples{
npars(as.term(c("sigma", "alpha[1]", "alpha[2]", "beta[1,1]", "beta[2,1]")))
npars(as.term(c("sigma", "alpha[1]", "alpha[2]")), scalar = TRUE)
}
\seealso{
\code{\link[base:dim]{base::dim()}}

Other {dimensions}: 
\code{\link[universals]{ndims}()},
\code{\link[universals]{npdims}()},
\code{\link[universals]{pdims}()}
}
