% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npdims.R
\name{npdims}
\alias{npdims}
\title{Number of Parameter Dimensions}
\usage{
npdims(x, ...)
}
\arguments{
\item{x}{The object}

\item{...}{Unused.}
}
\value{
A named list of the number of dimensions of each parameter.
}
\description{
Gets the number of parameter dimensions of an object as returned by \code{\link{pdims}()}.
}
\examples{
npdims(as.term(c("alpha[1]", "alpha[3]", "beta[1,1]", "beta[2,1]")))
}
\seealso{
\code{\link{pdims}()}
}
