% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npdims.R
\name{npdims.term}
\alias{npdims.term}
\title{Number of Dimensions of Each Parameter}
\usage{
\method{npdims}{term}(x, terms = FALSE, ...)
}
\arguments{
\item{x}{An object.}

\item{terms}{A flag specifying whether to get the number of dimensions for each term element.}

\item{...}{Other arguments passed to methods.}
}
\value{
A named integer vector of the number of dimensions of each parameter.
}
\description{
The terms argument is  \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}.
}
\examples{
npdims(term("alpha[1]", "alpha[3]", "beta[1,1]", "beta[2,1]"))
}
\seealso{
Other dimensions: 
\code{\link[universals]{dims}()},
\code{\link[universals]{ndims}()},
\code{\link[universals]{pdims}()}
}
