% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nterms.R
\name{nterms}
\alias{nterms}
\title{Number of Terms}
\usage{
nterms(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Unused.}
}
\value{
A count of the number of terms.
}
\description{
Gets the number of terms of an object.
}
\examples{
nterms(c("alpha[1]", "alpha[2]", "beta[1,1]", "beta[2,1]"))
}
