% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nterms.R
\name{nterms}
\alias{nterms}
\alias{nterms.term}
\title{Number of Terms}
\usage{
nterms(x, ...)

\method{nterms}{term}(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Unused.}
}
\value{
A count of the number of terms.
}
\description{
Gets the number of unique term elements of an object.
}
\details{
By default if the vector includes
missing values then it returns NA_integer_.
}
\section{Methods (by class)}{
\itemize{
\item \code{term}: Number of unique term elements of term vector
}}

\examples{
nterms(as.term(c("alpha[1]", "alpha[2]", "beta[1,1]", "beta[2,1]")))
nterms(as.term(c("alpha[1]", "alpha[1]", "beta[1,1]", "beta[1,1]")))
}
\seealso{
\code{\link[=term-vector]{term-vector()}}
}
