% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Parameter Descriptions for term Functions}
\arguments{
\item{x}{The object.}

\item{name}{A string of the name for the parameter.}

\item{...}{Unused.}

\item{repair}{A flag specifying whether to repair terms.

npdims}

\item{terms}{A flag specifying whether to get the number of dimensions for each
term element.}
}
\description{
as-term, is, complete-terms, consistent-term
}
