% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Parameter Descriptions for term Functions}
\arguments{
\item{x}{The object.}

\item{name}{A string specifying the name of the parameter.}

\item{...}{Unused.}

\item{select}{A character vector of the names of the parameters to include in the subsetted object.}

\item{value}{A character vector of the new parameter names.}

\item{pars}{A character vector of the new parameter names.}

\item{scalar}{A logical scalar specifying whether to include
all parameters (NA), only scalars (TRUE) or all parameters
except scalars (FALSE).}

\item{repair}{A logical scalar specifying whether to repair the term vector.}

\item{incomparables}{Ignored.}
}
\description{
Parameter Descriptions for term Functions
}
