% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Parameter Descriptions for term Functions}
\arguments{
\item{x}{The object.}

\item{name}{A string specifying the name of the parameter.}

\item{...}{Unused.}

\item{repair}{A flag specifying whether to repair terms.}

\item{select}{A character vector of the names of the parameters to include in the subsetted object.}

\item{value}{A character vector of the new parameter names.}

\item{pars}{A character vector of the new parameter names.}

\item{incomparables}{Ignored.}
}
\description{
Parameter Descriptions for term Functions
}
