% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pars.R
\name{pars.character}
\alias{pars.character}
\title{Parameter Names}
\usage{
\method{pars}{character}(x, scalar = NULL, ...)
}
\arguments{
\item{x}{An object.}

\item{scalar}{A flag specifying whether to by default return all parameters (NULL), or only scalar parameters (TRUE) or only non-scalar parameters (FALSE).}

\item{...}{Other arguments passed to methods.}
}
\value{
A character vector of the names of the parameters.
}
\description{
Gets the parameter names.
}
\examples{
pars(c("a", "b[1]", "a[3]"))
}
\seealso{
\link[universals:pars]{universals::pars}

Other parameters: 
\code{\link{pars.default}()},
\code{\link{pars.term}()},
\code{\link{pars.term_rcrd}()},
\code{\link{pars_terms}()}
}
\concept{parameters}
