% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pars.R
\name{pars.default}
\alias{pars.default}
\title{Parameter Names}
\usage{
\method{pars}{default}(x, scalar = NULL, ...)
}
\arguments{
\item{x}{An object.}

\item{scalar}{A flag specifying whether to by default return all parameters (NULL), or only scalar parameters (TRUE) or only non-scalar parameters (FALSE).}

\item{...}{Other arguments passed to methods.}
}
\value{
A character vector of the names of the parameters.
}
\description{
Gets the parameter names.
}
\examples{
pars(matrix(1:4, nrow = 2))
}
\seealso{
Other parameters: 
\code{\link[universals]{npars}()},
\code{\link[universals]{set_pars}()}
}
