% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/pars.R
\name{parameters}
\alias{parameters}
\alias{parameters<-}
\alias{set_parameters}
\alias{pars}
\alias{pars<-}
\alias{set_pars}
\alias{pars.term}
\title{Parameter Names}
\usage{
parameters(x, ...)

parameters(x) <- value

set_parameters(x, parameters)

pars(x, ...)

pars(x) <- value

set_pars(x, pars)

\method{pars}{term}(x, scalar_only = FALSE, terms = FALSE, ...)
}
\arguments{
\item{x}{An R object.}

\item{...}{Unused.}

\item{value}{A character vector of the new parameter names.}

\item{parameters}{A character vector of the new parameter names.}

\item{pars}{A character vector of the new parameter names.}

\item{scalar_only}{A flag specifying whether to only get the names of
parameters with one term.}

\item{terms}{A flag specifying whether to return the parameter name
for each term.}
}
\value{
A character vector of the parameter names.
}
\description{
Gets or sets the parameter names for an object.
}
\section{Methods (by class)}{
\itemize{
\item \code{term}: Parameter names for a term vector
}}

\examples{
term <- as.term(c("alpha[1]", "alpha[2]", "beta[1,1]", "beta[2,1]",
"beta[1,2]", "beta[2,2]", "sigma"))
pars(term)
pars(term, scalar_only = TRUE)
pars(term, terms = TRUE)
}
