% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/pars.R
\name{parameters}
\alias{parameters}
\alias{pararameters<-}
\alias{set_parameters}
\alias{pars}
\alias{pars<-}
\alias{set_pars}
\alias{pars.term}
\title{Parameter Names}
\usage{
parameters(x, ...)

pararameters(x) <- value

set_parameters(x, pars)

pars(x, ...)

pars(x) <- value

set_pars(x, pars)

\method{pars}{term}(x, scalar = NA, terms = FALSE, ...)
}
\arguments{
\item{x}{An R object.}

\item{...}{Unused.}

\item{value}{A character vector of the new parameter names.}

\item{pars}{A character vector of the new parameter names.}

\item{scalar}{A logical scalar specifying whether to get the names of
all parameters (NA), only scalars (TRUE) or all parameters
except scalars (FALSE).}

\item{terms}{A flag specifying whether to return the parameter name
for each term element.}
}
\value{
A character vector of the parameter names.
}
\description{
Gets or sets the parameter names for an object.
}
\details{
\code{parameters()} is deprecated.

\verb{parameters<-} is deprecated.

\code{set_parameters()} is deprecated.
}
\section{Methods (by class)}{
\itemize{
\item \code{term}: Parameter names for a term vector
}}

\examples{
term <- as.term(c(
  "alpha[1]", "alpha[2]", "beta[1,1]", "beta[2,1]",
  "beta[1,2]", "beta[2,2]", "sigma", NA
))
pars(term)
pars(term, scalar = TRUE)
pars(term, terms = TRUE)
}
