% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pars.R
\name{pars.term}
\alias{pars.term}
\title{Parameter Names}
\usage{
\method{pars}{term}(x, scalar = NA, terms = FALSE, ...)
}
\arguments{
\item{x}{An object.}

\item{scalar}{A logical scalar specifying whether to include
all parameters (NA), only scalars (TRUE) or all parameters
except scalars (FALSE).}

\item{terms}{A flag specifying whether to return the parameter name
for each term element.}

\item{...}{Other arguments passed to methods.}
}
\value{
A character vector of the parameter names.
}
\description{
Gets the parameter names.
}
\examples{
term <- as.term(c(
  "alpha[1]", "alpha[2]", "beta[1,1]", "beta[2,1]",
  "beta[1,2]", "beta[2,2]", "sigma", NA
))
pars(term)
}
\seealso{
Other {parameters}: 
\code{\link{pars_scalar}()},
\code{\link{pars_terms}()}
}
\concept{{parameters}}
