% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pars-terms.R
\name{pars_terms}
\alias{pars_terms}
\title{Term Parameters}
\usage{
pars_terms(x, scalar = NULL, ...)
}
\arguments{
\item{x}{A term vector.}

\item{scalar}{A flag specifying whether to by default return all parameters (NULL), or only scalar parameters (TRUE) or only non-scalar parameters (FALSE).}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
A character vector of the term parameter names.
}
\description{
Gets the name of each parameter for each term.
}
\details{
The scalar argument is \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}.
}
\examples{
term <- term(
  "alpha[1]", "alpha[2]", "beta[1,1]", "beta[2,1]",
  "beta[1,2]", "beta[2,2]", "sigma", NA
)
pars_terms(term)
}
\seealso{
Other parameters: 
\code{\link{pars.character}()},
\code{\link{pars.default}()},
\code{\link{pars.term}()},
\code{\link{pars.term_rcrd}()}
}
\concept{parameters}
