% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pars-terms.R
\name{pars_terms.term}
\alias{pars_terms.term}
\title{Get Term Parameters}
\usage{
\method{pars_terms}{term}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed to methods.}
}
\value{
A character vector of the parameter names.
}
\description{
Gets a character vector of the name of each parameter for each term.
}
\examples{
term <- as.term(c(
  "alpha[1]", "alpha[2]", "beta[1,1]", "beta[2,1]",
  "beta[1,2]", "beta[2,2]", "sigma", NA
))
pars_terms(term)
}
\seealso{
Other {parameters}: 
\code{\link[universals]{npars}()},
\code{\link[universals]{pars}()},
\code{\link[universals]{set_pars}()}

Other {terms}: 
\code{\link[universals]{complete_terms}()},
\code{\link[universals]{nterms}()}
}
