% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dims.R, R/npars.R, R/npdims.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{dims}
\alias{npars}
\alias{npdims}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{universals}{\code{\link[universals]{dims}}, \code{\link[universals]{npars}}, \code{\link[universals]{npdims}}}
}}

