% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dims.R, R/ndims.R, R/npars.R, R/npdims.R,
%   R/nterms.R, R/pars.R, R/pdims.R, R/set-pars.R, R/vec-cast.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{dims}
\alias{ndims}
\alias{npars}
\alias{npdims}
\alias{nterms}
\alias{pars}
\alias{pdims}
\alias{set_pars}
\alias{vec_cast}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{universals}{\code{\link[universals]{dims}}, \code{\link[universals]{ndims}}, \code{\link[universals]{npars}}, \code{\link[universals]{npdims}}, \code{\link[universals]{nterms}}, \code{\link[universals]{pars}}, \code{\link[universals]{pdims}}, \code{\link[universals]{set_pars}}}

  \item{vctrs}{\code{\link[vctrs]{vec_cast}}}
}}

