% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repair-terms.R
\name{repair_terms}
\alias{repair_terms}
\title{Repair Terms}
\usage{
repair_terms(x, scalars = TRUE)
}
\arguments{
\item{x}{A term vector.}

\item{scalars}{A flag specifying whether to repair scalars (see details).}
}
\value{
The repaired term vector.
}
\description{
Repairs a terms vector. I
}
\details{
Invalid elements are replaced by missing values and spaces removed.

By default (\code{scalars = TRUE}) if a parameter such as \code{b} 
is a scalar then \code{b[1]} is replaced by \code{b}
but if higher indices are included such as \code{b[2]} then \code{b}
is replaced by \code{b[1]}.
}
\examples{
repair_terms(as.term(c("b[3]", "b")))
repair_terms(as.term(c("a[3]", "b[1]")))
repair_terms(as.term(c("a [3]", " b [ 1  ] ")))
repair_terms(as.term(c("a", NA)))
}
\seealso{
\code{\link{term-vector}} and \code{\link{valid_term}}
}
