% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalar-term.R
\name{scalar_term}
\alias{scalar_term}
\title{Scalar Terms}
\usage{
scalar_term(x, repair = TRUE)
}
\arguments{
\item{x}{The object.}

\item{repair}{A logical scalar specifying whether to repair the term vector.}
}
\value{
A logical vector indicating whether each term is scalar.
}
\description{
Test whether each element in a term vector is scalar.
}
\details{
The function tests for the absence of a \code{[}
after by default repairing the term vector.
}
\examples{
scalar_term(as.term(c("a", "a[2]", "b[1]", "c[1,1]")))
}
\seealso{
\code{\link[=term-vector]{term-vector()}} and \code{\link[=repair_terms]{repair_terms()}}
}
