% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalar-term.R
\name{scalar_term}
\alias{scalar_term}
\title{Scalar Term}
\usage{
scalar_term(x)
}
\arguments{
\item{x}{The object.}
}
\value{
A logical vector indicating whether the term is a scalar.
}
\description{
Test whether each term is a scalar.
}
\examples{
scalar_term(term("alpha[1]", "alpha[3]", "beta[1]", "sigma[3]"))
scalar_term(term("alpha[1]", NA_term_, "beta[1]", "beta[3]"))
}
