% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset.term}
\alias{subset.term}
\title{Subset term vector}
\usage{
\method{subset}{term}(x, pars = NULL, ...)
}
\arguments{
\item{x}{A term vector to be subsetted.}

\item{pars}{A character vector of the parameters to include in the subsetted vector.}

\item{...}{Unused.}
}
\value{
The modified term vector.
}
\description{
Subset term vector
}
\examples{
term <- as.term(c("alpha[1]", "alpha[2]", "beta[1,1]", "beta[2,1]",
"beta[1,2]", "beta[2,2]", "sigma"))
subset(term, "beta")
subset(term, c("alpha", "sigma"))
}
