% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset.term}
\alias{subset.term}
\title{Subset Term Vector}
\usage{
\method{subset}{term}(x, select = NULL, ...)
}
\arguments{
\item{x}{A term vector to be subsetted.}

\item{select}{A character vector of the names of the parameters to include in the subsetted object.}

\item{...}{Unused.}
}
\value{
The modified term vector.
}
\description{
Subsets a term vector.
}
\examples{
term <- as.term(c(
  "alpha[1]", "alpha[2]", "beta[1,1]", "beta[2,1]",
  "beta[1,2]", "beta[2,2]", "sigma"
))
subset(term, "beta")
subset(term, c("alpha", "sigma"))
}
\seealso{
\code{\link[=term-vector]{term-vector()}}
}
