% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term.R
\name{term}
\alias{term}
\alias{term-object}
\alias{term_object}
\alias{term-vector}
\alias{term_vector}
\alias{term.integer}
\title{Term Vector}
\usage{
term(x, ...)

\method{term}{integer}(x, name = "par", ...)
}
\arguments{
\item{x}{An integer vector of the object dimensions or a named list of the parameter's dimensions.}

\item{...}{Unused.}

\item{name}{A string specifying the name of the parameter.}
}
\value{
A term vector.
}
\description{
Creates a term vector from object dimensions.
A \code{term} vector is a S3 vector of parameter terms.
}
\section{Methods (by class)}{
\itemize{
\item \code{integer}: Term vector from an integer vector of the object's dimensions
}}

\examples{
term(0L)
term(1L)
term(2L)
}
\seealso{
\code{\link[=dims]{dims()}} and \code{\link[=pdims]{pdims()}}
}
