% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term.R
\name{term}
\alias{term}
\title{Object Terms}
\usage{
term(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{...}{Unused.}
}
\value{
A term vector.
}
\description{
Gets term vector for an R object.
}
\examples{
term(c("alpha[1]", "alpha[2]", "beta[1,1]", "beta[2,1]",
"beta[1,2]", "beta[2,2]", "sigma"))
}
