% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term.R
\name{term}
\alias{term}
\title{Term Vectors}
\usage{
term(dims = 0L, name = "par")
}
\arguments{
\item{dims}{An integer vector of the parameter dimensions.}

\item{name}{A string specifying the name of the parameter 
for vectors of positive length.}
}
\value{
A term vector.
}
\description{
Creates a term vector.
}
\examples{
term(0L)
term(1L)
term(2L)
}
\seealso{
\code{\link{term-vector}} and \code{\link{dims}()}
}
