% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tindex.R
\name{tindex}
\alias{tindex}
\title{Term Index}
\usage{
tindex(x, ...)
}
\arguments{
\item{x}{The object}

\item{...}{Unused.}
}
\description{
Gets a list of the index for each term of an object.
}
\details{
\code{tdims()} is deprecated.
}
\examples{
tindex(c("alpha[1]", "alpha[2]", "beta[1,1]", "beta[2,1]", 
"beta[1,2]", "beta[2,2]", "sigma"))
}
