% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique.R
\name{unique.term}
\alias{unique.term}
\title{Extract Unique Elements}
\usage{
\method{unique}{term}(x, incomparables = FALSE, ...)
}
\arguments{
\item{x}{A term vector.}

\item{incomparables}{Ignored.}

\item{...}{Unused.}
}
\value{
A term vector.
}
\description{
Extract Unique Elements
}
\examples{
unique(term(c("a[2]", "a[1]", "a[1]")))
}
