% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique.R
\name{unique.term}
\alias{unique.term}
\title{Extract Unique Terms}
\usage{
\method{unique}{term}(x, incomparables = FALSE, ...)
}
\arguments{
\item{x}{A term vector.}

\item{incomparables}{Ignored.}

\item{...}{Unused.}
}
\value{
A vector of unique terms.
}
\description{
Extracts unique elements from a term vector.
}
\examples{
unique(as.term(c("b", "a", "a[1]", "a", NA, "aa", NA)))
}
\seealso{
\code{\link[=term-vector]{term-vector()}}
}
