% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec-ptype2.R
\name{vec_ptype2.term}
\alias{vec_ptype2.term}
\title{Find the common type for a pair of vectors}
\usage{
\method{vec_ptype2}{term}(x, y, ...)
}
\arguments{
\item{x}{Vector types.}

\item{y}{Vector types.}

\item{...}{These dots are for future extensions and must be empty.}
}
\description{
\code{vec_ptype2()} defines the coercion hierarchy for a set of related
vector types. Along with \code{\link[vctrs:vec_cast]{vec_cast()}}, this generic forms the
foundation of type coercions in vctrs.

\code{vec_ptype2()} is relevant when you are implementing vctrs methods
for your class, but it should not usually be called directly. If
you need to find the common type of a set of inputs, call
\code{\link[vctrs:vec_ptype]{vec_ptype_common()}} instead. This function supports multiple
inputs and \link[vctrs:new_partial]{finalises} the common type.
}
\seealso{
\code{\link[vctrs:vec_ptype2]{vctrs::vec_ptype2()}}
}
